Only asks for creds if request is unauthorized.
For DockerHub specifically, even public data requires an auth token, but
this does not require credentials to get.
So, on the first request we send empty creds, and if there is an auth
error we'll try again with real creds.

Push is a little more tricky, so we configure the resolver to send the
creds even the 1st time they are asked for, which is probably required
everywhere anyway.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>